Bugfixing 1

+++ WMSApp/WMSWebApp/Controllers/CompaniesController.cs
@@ -12,26 +12,36 @@ using Microsoft.AspNetCore.Mvc;
 using Newtonsoft.Json;
 using WMSWebApp.ViewModels;
 using Microsoft.Extensions.Configuration;
+using Domain.Model.CompanyMaster;
+using System.Linq;
+using Application.Services.Security;
+
 namespace WMSWebApp.Controllers
 {
     [Authorize]
-    public class CompaniesController : Controller
+    public class CompaniesController : BaseAdminController
     {
         private readonly ICompanyHelper _companyHelper;
         private readonly IItemHelper _itemHelper;
         private readonly IMapper _mapper;
         private Microsoft.Extensions.Configuration.IConfiguration Configuration;
-        public CompaniesController(ICompanyHelper companyHelper, IMapper mapper, IItemHelper itemHelper, Microsoft.Extensions.Configuration.IConfiguration _configuration)
+        private readonly IPermissionMasterService _permissionMasterService;
+        public CompaniesController(ICompanyHelper companyHelper, IMapper mapper, IItemHelper itemHelper, Microsoft.Extensions.Configuration.IConfiguration _configuration, IPermissionMasterService permissionMasterService)
         {
             _companyHelper = companyHelper;
             _mapper = mapper;
             _itemHelper = itemHelper;
             Configuration = _configuration;
+            _permissionMasterService = permissionMasterService;
         }
 
         // GET: CompaniesController
         public ActionResult Index()
         {
+            if (!_permissionMasterService.Authorize(StandardPermissionProvider.Company, PermissionType.View).Result)
+            {
+                return AccessDeniedView();
+            }
             List<Company> companies = new List<Company>();
             try
             {
@@ -46,29 +56,33 @@ namespace WMSWebApp.Controllers
         }
 
         // GET: CompaniesController/Details/5
-        [NonAction]
-        public ActionResult Details(int id)
-        {
-            var c = new Company()
-            {
-                CompanyName = "Test",
-                CompanyCode = "Test",
-                Address = "Test",
-                Location = "Location",
-                ContactPersonHO = "ContactPersonHO",
-                ContactNumberHO = "ContactNumberHO"
-            ,
-                EmailIdHO = "EmailIdHO",
-                SpaceSizeFormat = "SpaceSizeFormat",
-                Items = 1,
-                SubItem = 1
-            };
-            return View(c);
-        }
+        //[NonAction]
+        //public ActionResult Details(int id)
+        //{
+        //    var c = new Company()
+        //    {
+        //        CompanyName = "Test",
+        //        CompanyCode = "Test",
+        //        Address = "Test",
+        //        Location = "Location",
+        //        ContactPersonHO = "ContactPersonHO",
+        //        ContactNumberHO = "ContactNumberHO"
+        //    ,
+        //        EmailIdHO = "EmailIdHO",
+        //        SpaceSizeFormat = "SpaceSizeFormat",
+        //        Items = 1,
+        //        SubItem = 1
+        //    };
+        //    return View(c);
+        //}
 
         // GET: CompaniesController/Create
         public ActionResult Create()
         {
+            if (!_permissionMasterService.Authorize(StandardPermissionProvider.Company, PermissionType.Add).Result)
+            {
+                return AccessDeniedView();
+            }
             Company company = new Company();
             var itemData = _itemHelper.GetAllItem();
             company.ItemList = _mapper.Map<List<Item>>(itemData);
@@ -92,6 +106,18 @@ namespace WMSWebApp.Controllers
             try
             {
                 var company = _mapper.Map<CompanyDb>(c);
+
+                foreach (var item in c.Items)
+                {
+                    CompanyItemsMapping itemMapping = new CompanyItemsMapping
+                    {
+                        Company = company,
+                        ItemId = item,
+                        CompanyId = company.Id
+                    };
+                    company.CompanyItemsMappings.Add(itemMapping);
+                }
+
                 _companyHelper.Insert(company);
                 return RedirectToAction(nameof(Index));
             }
@@ -104,11 +130,19 @@ namespace WMSWebApp.Controllers
         // GET: CompaniesController/Edit/5
         public ActionResult Edit(int id)
         {
+            if (!_permissionMasterService.Authorize(StandardPermissionProvider.Company, PermissionType.Edit).Result)
+            {
+                return AccessDeniedView();
+            }
             var c = new Company();
             try
             {
-                var data = _companyHelper.GetCompanyById(id);
+               
+                var data = _companyHelper.GetById(id);
                 c = _mapper.Map<Company>(data);
+                var itemData = _itemHelper.GetAllItem();
+                c.ItemList = _mapper.Map<List<Item>>(itemData);
+                c.Items = data.CompanyItemsMappings.ToList().Select(x => x.ItemId).ToList();
             }
             catch (Exception)
             {
@@ -137,6 +171,10 @@ namespace WMSWebApp.Controllers
         // GET: CompaniesController/Delete/5
         public ActionResult Delete(int id)
         {
+            if (!_permissionMasterService.Authorize(StandardPermissionProvider.Company, PermissionType.Delete).Result)
+            {
+                return AccessDeniedView();
+            }
             try
             {
                 var b = _companyHelper.DeleteCompanyById(id);
